Avoid re-creating functions on every render
Previously, the project created anonymous or new functions via bind
inside of several render methods. This is hard for React to optimize
since the function references will change on every tick of the component
life-cycle i.e, React will have to assume that the component changed and
re-render it on every cycle which leads to poor performance.

One clean way to solve the this-context and anonymous functions problem
at the same time is to use methods at the class level which will
automatically preserve the correct this-context while also being
constant method references i.e., easy for React to optimize.

You can learn more about this approach here:
https://medium.freecodecamp.org/react-binding-patterns-5-approaches-for-handling-this-92c651b5af56